% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{accumulation.thresholds}
\alias{accumulation.thresholds}
\title{Title}
\usage{
accumulation.thresholds(x, boost = 12, max = 76)
}
\arguments{
\item{x}{}

\item{boost}{}

\item{max}{}
}
\description{
Title
}
