% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_density}
\alias{add_density}
\title{Density or height curves for graphs}
\usage{
add_density(layout, ...)
}
\arguments{
\item{layout}{a matrix with coordinates often produced by \link{layout_with_fr}}

\item{...}{}
}
\value{

}
\description{
Height curves show where on a plot the concentration of points is largest.
In the example below we are interested in where in the network women are concentrated.
We compare the density curves of the women with that of the men and we see that men are slightly more concentrated in the core of the network.
}
\examples{
data(den)
den.health     <- has.tags(den, "Health", res = "den")
graph.health   <- elite.network(den.health)
graph.health   <- largest.component(graph.health)
women          <- V(graph.health)$name \%in\% den.health$NAME[den.health$GENDER == "Women"]
lay.health     <- layout_with_fr(graph.health, grid = "nogrid")
p              <- graph.plot(graph.health, lay.health, vertex.fill = women, edge.color = "black", edge.size = 0.3)
p              <- p + scale_fill_manual(values = c("white", "black"), name = "Women")
p + add_density(lay.health[women,], color = "darkred")
p + add_density(lay.health[women == FALSE,], color = "darkred")
}
