% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph creation.R
\name{adj.affil}
\alias{adj.affil}
\title{Adjacency matrix for affiliations}
\usage{
adj.affil(den)
}
\arguments{
\item{den}{an affiliation list}
}
\value{
a sparse adjacency matrix of affiliation * affiliation
}
\description{
Create an adjacency matrix from an affilation list
}
\seealso{
\link{adj.ind}, \link{two.mode}
}

