% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{as.den}
\alias{as.den}
\title{Convert data.frame to the den format.}
\usage{
as.den(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
a den-class object
}
\description{
If you have a data.frame with case - affiliation edges: where each row is a position in a affiliation, it can be converted to den class object.
The required column names are "NAME" for the names of the individuals and "AFFILIATION" for the names of the affiliations.
Common variables include: ROLE, TAGS, PERSON_ID, DESCRIPTION, GENDER
}
\examples{
data(den)
as.den(den)
}

