% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core identification.R
\name{betweenness.decomposition}
\alias{betweenness.decomposition}
\title{Betweenness decomposition}
\usage{
betweenness.decomposition(
  graph,
  max.path = 2,
  estimate.min = 1,
  silent = FALSE
)
}
\arguments{
\item{graph}{a network graph}

\item{max.path}{the longest paths used by \link{betweenness.estimate}}

\item{estimate.min}{}
}
\value{
a graph
}
\description{
Find the people that link the network together. All individuals with less than the minimum betweenness are removed from the network.
This removes "hangers" and people with connections that are perfectly similar to others. Betweenness.decomposition is useful before core identification and as a pruning tool for "dirty" data.
If you choose a relatively high minimum betweenness estimate then it can serve as a core identification method.
}
\examples{
data(pe13)
betweenness.decomposition(net.elite)
}
