% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{code.gender}
\alias{code.gender}
\title{Code gender by firstname in Denmark}
\usage{
code.gender(x)
}
\arguments{
\item{x}{a character vector of full names}
}
\value{
a factor with a gender guess
}
\description{
Guesses the gender for a list of names by comparing it to the Danish national distribution of first names in 2013.
}
\examples{
data(den)
priest.names   <- has.tags(den, "Churches", result = "name")
table(code.gender(priest.names))
}
