% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{code.region}
\alias{code.region}
\title{Categories from postal codes}
\usage{
code.region(x)
}
\arguments{
\item{x}{a numeric vector with 4 digit danish postal codes}
}
\value{
a data.frame with various factors
}
\description{
Categories from postal codes
}

