% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{description.vector}
\alias{description.vector}
\title{A vector of descriptions for each individual}
\usage{
description.vector(x, den)
}
\arguments{
\item{x}{a character vector with names of individuals}

\item{den}{a den class object with a DESCRIPTION variable, see \link{as.den}}
}
\value{
a vector with descriptions where affiliations are seperated by "*"
}
\description{
A vector of descriptions for each individual
}
