% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynacores.R
\name{dynamic.core.analysis}
\alias{dynamic.core.analysis}
\title{Identify dynamic cores}
\usage{
dynamic.core.analysis(
  den,
  graph.spell = NULL,
  adjacency.list = NULL,
  period = c(ymd(19880101), ymd(19901231)),
  median.term = 38,
  max.terms = 2,
  boost = 38
)
}
\arguments{
\item{den}{a den class object}

\item{graph.spell}{a graph.spell - if NULL it is calculated}

\item{adjacency.list}{a list of weighted adjacency matrices - if NULL it is calculated}

\item{period}{a vector with start date and end date, see \link{ymd}}

\item{median.term}{the median number of months a term lasts}

\item{max.terms}{the maximum number of terms}

\item{boost}{the minimum value of active relations}
}
\value{

}
\description{
Identify dynamic cores
}
