% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph creation.R
\name{elite.network.affil}
\alias{elite.network.affil}
\title{Elite network for affiliations}
\usage{
elite.network.affil(den = den, sigma = 14)
}
\arguments{
\item{den}{an affiliation edge list in the \link{as.den} format.}

\item{sigma}{the number of members in an affiliation above which all affiliations are weighted down}
}
\value{
a elite network object
}
\description{
Construct a weighted elite network of affiliations
}

