% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph creation.R
\name{elite.network}
\alias{elite.network}
\title{Elite network}
\usage{
elite.network(den, sigma = 14)
}
\arguments{
\item{den}{an affiliation edge list, in the \link{as.den} format.}

\item{sigma}{the number of members in an affiliation above which all affiliations are weighted down}
}
\value{
a weighted igraph graph
}
\description{
Construct a weighted elite network
}
\examples{
data(den)
den.health <- has.tags(den, "Health", result = "den")
elite.network(den.health)
}

