% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph creation.R
\name{elite.network.two.mode}
\alias{elite.network.two.mode}
\title{A weighted two mode network}
\usage{
elite.network.two.mode(den, sigma = 14)
}
\arguments{
\item{den}{a den class object}

\item{sigma}{the number of members in an affiliation above which all affiliations are weighted down}
}
\value{
an igraph graph
}
\description{
A weighted two mode network
}
