% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\encoding{utf-8}
\name{eliteDB_close}
\alias{eliteDB_close}
\title{eliteDB_close}
\usage{
eliteDB_close(conn)
}
\arguments{
\item{conn}{en åben forbindelse til eliteDB}
}
\description{
Luk en aktiv forbindelse til eliteDB
}
\examples{
conn <- eliteDB_open(user, pass)
eliteDB_list_tables(conn)
persons <- eliteDB_read_table(conn,"Persons")
eliteDB_close(conn)
}
\seealso{
\code{\link{eliteDB_open}} 
    \code{\link{eliteDB_list_tables}}
     \code{\link{eliteDB_read_table}}
}
