% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\encoding{utf-8}
\name{eliteDB_cvr}
\alias{eliteDB_cvr}
\title{eliteDB_cvr}
\usage{
eliteDB_cvr(conn = NULL, user = NULL, pass = NULL,
  db = "elitedb_bigdata", cvr = NULL, ownership = NULL,
  employees = NULL)
}
\arguments{
\item{conn}{en åben forbindelse til eliteDB}

\item{user}{brugernavn}

\item{pass}{kodeord}

\item{db}{navn på den database data skal hentes fra. enten "elitedb_cvr" eller "elitedb_bigdata"}
}
\value{
Et komplet den-objekt, der både indeholder aktive og arkiverede forbindelser
}
\description{
Hent et komplet den-objekt fra elitedb_cvr
}
\examples{
den        <- eliteDB_cvr(user = user, pass = password, cvr = c(10014107, 10013569))
den        <- eliteDB_cvr(user = user, pass = password, cvr = c(10014107, 10013569), ownership = TRUE)
den        <- eliteDB_cvr(user = user, pass = password, cvr = c(10014107, 10013569), employees = TRUE)
}
