% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\encoding{utf-8}
\name{eliteDB_den}
\alias{eliteDB_den}
\title{eliteDB_den}
\usage{
eliteDB_den(conn = NULL, user = NULL, pass = NULL)
}
\arguments{
\item{conn}{en åben forbindelse til eliteDB}

\item{user}{brugernavn}

\item{pass}{kodeord}
}
\value{
Et komplet den-objekt, der både indeholder aktive og arkiverede forbindelser
}
\description{
Hent et komplet den-objekt fra eliteDB
}
\examples{
den        <- eliteDB_den(user = user, pass = password)
den.active <- den[is.na(den$ARCHIVED),]
}
