% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\encoding{utf-8}
\name{eliteDB_list_tables}
\alias{eliteDB_list_tables}
\title{eliteDB_list_tables}
\usage{
eliteDB_list_tables(conn)
}
\arguments{
\item{conn}{en åben forbindelse til eliteDB}
}
\description{
Se hvilke tabeler der er tilgængelige i eliteDB
}
\examples{
conn <- eliteDB_open(user, pass)
eliteDB_list_tables(conn)
persons <- eliteDB_read_table(conn,"Persons")
eliteDB_close(conn)
}
\seealso{
\code{\link{eliteDB_open}} 
    \code{\link{eliteDB_close}} 
    \code{\link{eliteDB_read_table}}
}
