% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\encoding{utf-8}
\name{eliteDB_open}
\alias{eliteDB_open}
\title{eliteDB_open}
\usage{
eliteDB_open(user, pass, db = "elitedb")
}
\arguments{
\item{user}{brugernavn}

\item{pass}{kodeord}

\item{db}{navnet på den database du ønsker at tilgå}
}
\value{
en aktiv MySQL forbindelse til eliteDB
}
\description{
Opret forbindelse til eliteDB
}
\details{
eliteDB_open bruges til at oprette en forbindelse til vores server, sådan at man kan begynde at hente data ned.
    Funktionen skal køres før man kan bruge de øvrige eliteDB funktioner (med undtagelse af eliteDB_den).
}
\examples{
conn <- eliteDB_open(user, pass)
eliteDB_list_tables(conn)
persons <- eliteDB_read_table(conn,"Persons")
eliteDB_close(conn)
}
\seealso{
\code{\link{eliteDB_close}} 
    \code{\link{eliteDB_list_tables}}
    \code{\link{eliteDB_read_table}}
}
