% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\encoding{utf-8}
\name{eliteDB_read_table}
\alias{eliteDB_read_table}
\title{eliteDB_read_table}
\usage{
eliteDB_read_table(conn, table)
}
\arguments{
\item{conn}{en åben forbindelse til eliteDB}

\item{table}{navn på den tabel der skal hentes.}
}
\description{
Hent data fra eliteDB
}
\examples{
conn <- eliteDB_open(user, pass)
eliteDB_list_tables(conn)
persons <- eliteDB_read_table(conn,"Persons")
eliteDB_close(conn)
}
\seealso{
\code{\link{eliteDB_open}} 
    \code{\link{eliteDB_close}}
    \code{\link{eliteDB_list_tables}}
}
