% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\encoding{utf-8}
\name{eliteDB_tag_rapport}
\alias{eliteDB_tag_rapport}
\title{eliteDB_tag_rapport}
\usage{
eliteDB_tag_rapport(conn)
}
\arguments{
\item{conn}{en åben forbindelse til eliteDB}
}
\description{
Få et overblik over tags i eliteDB. Funktionen returnerer en data.frame,
   der viser hvor mange affiliations, aktive forbindelser og arkiverede forbindelser, 
   der er knyttet til et tag.
}
\examples{
conn <- eliteDB_open(user, pass)
tags <- eliteDB_tag_rapport(conn)
}

