% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliter.R
\docType{package}
\name{eliter-package}
\alias{eliter}
\alias{eliter-package}
\title{eliter: Contains the Danish Elite Network for 2012 and 2017 with analytical functions.}
\description{
Here we collect and develop a set of tools for the analysis of affiliation networks. Currently it contains functions for weighting networks, temporal affiliation networks, k-shell decomposition, betweenness decomposition and a ton of convience and plotting functions. Furthermore this package contains data and code from published works on elites.
}
