% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{extract.postal.codes}
\alias{extract.postal.codes}
\title{Extract Danish postal codes from an adress string}
\usage{
extract.postal.codes(x, check.valid = FALSE)
}
\arguments{
\item{x}{a character vector with addresses}

\item{check.valid}{if TRUE all invalid postal codes are set as NA. This affects postal codes for Greenland (3900:3992)}
}
\value{
a numeric vector
}
\description{
Extract Danish postal codes from an adress string
}
\examples{
data(corp13)
address            <- corp13$Adress
plot(table(extract.postal.codes(address)))
}

