% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{from.spell.graph.to.survival}
\alias{from.spell.graph.to.survival}
\title{Turn a spell.graph into a dataset suitable for survival analysis}
\usage{
from.spell.graph.to.survival(spell.graph, end.month = 1392)
}
\arguments{
\item{spell.graph}{}

\item{end.month}{an integer for the end month. The end month is the end point for all non-reoccuring ties that are broken.}
}
\value{
a data.frame
}
\description{
Note that this code is slow.
}
