% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{geographical.neighbours}
\alias{geographical.neighbours}
\title{Geoggraphical neighbours graph}
\usage{
geographical.neighbours(lon, lat, id = 1:length(lon), distance = 500)
}
\arguments{
\item{lon}{a numeric vector with longitudes}

\item{lat}{a numeric vector with latitudes}

\item{id}{a character vector with names}

\item{distance}{maximum distance in meters}
}
\value{
a weighted graph with edge weights equal to the geographical distance
}
\description{
How many people do you live close to within a given distance in meters
}
\examples{
data(pe13)
graph.neighbour       <- geographical.neighbours(lon = pe13$lon, lat = pe13$lat, id = pe13$Name, distance = 500)
graph.plot(graph.neighbour)
}

