% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynacores.R
\name{get.shell}
\alias{get.shell}
\title{Get shell from adjacency matrix}
\usage{
get.shell(x, start.level = 20, verbose = FALSE)
}
\arguments{
\item{x}{an sparse adjacency matrix}

\item{start.level}{the first level of the k-shell decomposition}

\item{verbose}{if TRUE prints a lot of info}
}
\value{

}
\description{
Get shell from adjacency matrix
}
