% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{has.tags}
\alias{has.tags}
\title{Has.tags}
\usage{
has.tags(den, tags, result = c("affil", "name", "den"), silent = FALSE)
}
\arguments{
\item{den}{an affiliation edge list}

\item{tags}{is a character vector of tags}

\item{result}{if "den" \code{has.tags} returns a character vector with unique affiliation names. If result is "name", a character vector of unique names is returned. If result is "den", a den object is returned.}

\item{silent}{if TRUE the table with the number of matched positions per tag is not shown.}
}
\description{
Selects all affiliations that have one or several tags
}
\examples{
data(den)
has.tags(den, tags = c("Youth", "Children"))
has.tags(den, tags = c("Youth", "Children"), res = "name")
has.tags(den, tags = c("Youth", "Children"), res = "den")
}

