% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core identification.R
\name{k.shell}
\alias{k.shell}
\title{K-shell decomposition}
\usage{
k.shell(graph, start.level = 0, verbose = FALSE)
}
\arguments{
\item{graph}{a graph from igraph}

\item{start.level}{the decomposition process starts at this level or minimum degree}

\item{verbose}{if TRUE shows the progress}
}
\value{
a numeric vector
}
\description{
K-shell decomposition
}
\examples{
data(den)
graph.all <- elite.network(den)
graph.com <- largest.component(graph.all)
sp        <- distances(graph.com)
sp[sp >= 2.1] <- 0
graph.reach     <- graph_from_adjacency_matrix(Matrix(sp), mode = "undirected", weighted = TRUE)
shell <- k.shell(graph.reach, start.level = 0, verbose = TRUE)
table(shell)
}
