% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core identification.R
\name{largest.component}
\alias{largest.component}
\title{Largest component}
\usage{
largest.component(graph, cut.off = 1, result = c("graph", "vector"))
}
\arguments{
\item{graph}{a \link{igraph} network object}

\item{cut.off}{is the minimum number of weighted affiliation memberships a component member has to have.}

\item{result}{either a graph with the largest component or a logical vector.}
}
\value{
a graph object
}
\description{
Finds the largest component in a network.
}
\examples{
data(den)
den.health   <- has.tags(den, "Health", res = "relations")
graph.health <- elite.network(den.health)
largest.component(graph.health)
}

