% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{lastnames}
\alias{lastnames}
\title{Extract last names}
\usage{
lastnames(x)
}
\arguments{
\item{x}{a character vector of full nmaes}
}
\value{
a character vector of last names
}
\description{
Extract last names from full names
}
\examples{
data(den)
last          <- table(lastnames(den$NAME))
head(sort(last, decreasing = TRUE), 20)
}
