% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout.cl}
\alias{layout.cl}
\title{Title}
\usage{
layout.cl(graph, decompose = TRUE, n = 3, remove.isolates = FALSE, ...)
}
\arguments{
\item{graph}{}

\item{...}{}
}
\description{
Title
}
\examples{
data(den)
den.culture   <- has.tags(den, "Culture", result = "den")
graph         <- elite.network(den.culture, result = "affil")
plot(layout.cl(graph, decompose = TRUE))
}

