% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-circles.R
\name{level.membership}
\alias{level.membership}
\title{Level membership from minimal membership decomposition}
\usage{
level.membership(
  l.inc,
  mode = c("ind", "affil", "two-mode"),
  levels = seq_along(l.inc)
)
}
\arguments{
\item{l.inc}{a list of nested incidence matrices}
}
\value{
a tibble with rownames and level membership
}
\description{
Level membership from minimal membership decomposition
}
