% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core identification.R
\name{levels.of.power}
\alias{levels.of.power}
\title{Levels of power}
\usage{
\method{levels}{of.power}(x)
}
\arguments{
\item{x}{a numeric vector with coreness scores produced by \link{find.core}}
}
\value{
a factor with 4 levels
}
\description{
Levels of power
}
