% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia match.R
\name{match_and_write_chunks_from_wiki}
\alias{match_and_write_chunks_from_wiki}
\title{Match search terms to Wikipedia}
\usage{
match_and_write_chunks_from_wiki(
  x,
  file = "saved/wiki_match.Rda",
  chunk.size = 10,
  save.every = 10,
  key.words = c("city"),
  blacklist = c("is a list of", "may also refer to", "commonly refers to"),
  valid.qids = NULL
)
}
\arguments{
\item{x}{a character vector}

\item{file}{a filepath}

\item{chunk.size}{the number of search terms in each chunk - it cannot be above 50}

\item{save.every}{the number of chunks between each save to file}

\item{key.words}{a character vector with the keywords that should be present in a wiki article or in the wiki categories}

\item{blacklist}{a character vector with blacklisted terms or phrases.}
}
\value{
a tibble
}
\description{
Match search terms to Wikipedia
}
