% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{member.vector}
\alias{member.vector}
\title{A vector of members in an affiliation}
\usage{
member.vector(x, den)
}
\arguments{
\item{x}{a character vector with names of affiliations}

\item{den}{}
}
\value{
a vector
}
\description{
A vector of members in an affiliation
}

