% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{members.vector}
\alias{members.vector}
\title{A vector of members in an affiliation}
\usage{
members.vector(x, den)
}
\arguments{
\item{x}{a character vector with names of affiliations}

\item{den}{a den class object with the affiliations in x}
}
\value{
a vector with names of the members seperated by " * "
}
\description{
A vector of members in an affiliation
}
\examples{
data(den)
den.culture   <- has.tags(den, "Culture", result = "den")
graph.culture <- elite.network(den.culture, result = "affil")
members.vector(V(graph.culture)$name, den)
}

