% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core identification.R
\name{minimal.members.decomposition}
\alias{minimal.members.decomposition}
\title{Minimual members decomposition}
\usage{
minimal.members.decomposition(incidence, minimum.memberships = 3)
}
\arguments{
\item{incidence}{a sparse incidence matrix}

\item{minimum.memberships}{the minimum number of memberships for the individuals (rows)}
}
\value{
a list of incidence matrices
}
\description{
Minimual members decomposition
}
\examples{
data(den)
den         <- den[den$SOURCE != "Events",]
incidence   <- xtabs(~NAME + AFFILIATION, droplevels(den), sparse = TRUE)
l.inc       <- minimal.members.decomposition(t(incidence), 3)
level.membership(l.inc)
l.inc[[5]] \%>\% colSums() \%>\% sort() \%>\% as.matrix()
l.inc[[5]] \%>\% rowSums() \%>\% sort() \%>\% as.matrix()
}
