% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{plot_cum_ind_pr_org}
\alias{plot_cum_ind_pr_org}
\title{Title}
\usage{
plot_cum_ind_pr_org(
  data,
  reference.year = 2013,
  cut.points = c(0.5, 0.75, 0.9, 0.95),
  x.position = 200
)
}
\arguments{
\item{data}{the sequence data set}
}
\value{

}
\description{
Title
}
\details{
data needs to have the following format: start and end in years. organization is the name of the organization.
}
\examples{
data("careers")
dt        <- sequence.data_raw[!gsub(",", ".", sequence.data_raw$`Career level`) \%in\% c(0.3, 0.5, 0.6, 0.7, 0.8,0.9),]
dt        <- data.frame(name = dt$NAME, organization = dt$Organisation, start = dt$Start, end = dt$End)
plot_cum_ind_pr_org(dt)
}
