% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.den}
\alias{print.den}
\title{Print den objects}
\usage{
\method{print}{den}(x, ...)
}
\arguments{
\item{...}{further arguments are ignored}

\item{den}{a den-class object}
}
\value{
a markdown table with statistics
}
\description{
Essential statistics for the den-class. Mainly for interactive use, but is returned in rmarkdown format using kable and knitr for easy export.
}

