% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{prior.connections}
\alias{prior.connections}
\title{Title}
\usage{
prior.connections(graph.spell, minimum.gap = 12)
}
\arguments{
\item{graph.spell}{}

\item{minimum.gap}{}
}
\value{

}
\description{
Title
}
