% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{reach}
\alias{reach}
\title{Reach centrality}
\usage{
reach(graph, order = 2, mode = "all", weight = E(graph)$weight,
  max.path = NULL)
}
\arguments{
\item{graph}{a weighted graph}

\item{order}{the maximum distance between two vertices}

\item{mode}{the direction of the edges; either, "in", "out" or "all"}

\item{weight}{the edge weight as distance}

\item{max.path}{the maximum unweighted path length}
}
\value{
a numeric vector with reach scores
}
\description{
The number of people you can reach within a few steps (usually 2). This is 
similar to the size of your neighborhood except that reach takes the strength
of the edges into account. The distance between two people is the sum of the
edge weights of the edges in the path between the two people. If you use the
max.path argument you can restrict the maximum number of people a "reach
path" can go through. This is relevant for networks with very strong
connections aka. short distances between agents.
}
\details{
Reach can be calculated for both directed and undirected graphs. For
unweighted graphs see \link{neighborhood.size} or \link{ego.size} from
\link{igraph}.
}
\examples{
data(pe13)
r <- reach(net.elite, order = 2)
tail(sort(r))
r <- reach(net.elite, order = 3, max.path = 2)
tail(sort(r))
}
\seealso{
ego, ego_size, degree, vertex.measures
}

