% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph creation.R
\name{reduced.sector.graph}
\alias{reduced.sector.graph}
\title{Create a reduced graph from a list of den objects}
\usage{
reduced.sector.graph(list.dens)
}
\arguments{
\item{list.dens}{a list of den objects, like those created by \link{tags.to.sectors}.}
}
\value{
an igraph object with additional attributes
}
\description{
Creates a reduced graph on the basis of odds-ratios. Note that if the den objects are not mutually exclusive it may influence the interpretation of the odds-ratio edges.
}
