% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{sector.connections}
\alias{sector.connections}
\title{Title}
\usage{
sector.connections(graph, you = V(graph)$name, sectors, how.well = 2)
}
\arguments{
\item{graph}{}

\item{you}{}

\item{sectors}{}

\item{how.well}{}
}
\description{
Title
}
