% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{sectors.to.role}
\alias{sectors.to.role}
\title{Title}
\usage{
sectors.to.role(den, list.dens, role = c("Chief executive", "Executive"),
  other = "Other", mutually.exclusive = FALSE)
}
\arguments{
\item{den}{}

\item{list.dens}{}

\item{role}{}

\item{other}{}

\item{mutually.exclusive}{}
}
\description{
Title
}
\examples{
data(den)
den                 <- as.den(den)
list.dens           <- tags.to.sectors(den, standard.sectors("Danish"))
sectors.to.role(den, list.dens, mutually.exclusive = TRUE)
}

