% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{standard.sectors}
\alias{standard.sectors}
\title{Standard sector tags}
\usage{
standard.sectors(sets = c("Danish", "English", "4 Sectors"))
}
\arguments{
\item{sets}{a character vector with the sets of sector tags}
}
\value{
a list of tags
}
\description{
Collections of tags that roughly correspond to sectors.
}
\examples{
data(den)
sectors      <- standard.sectors()
sectors$Danish

standard.sectors("English")
}
