% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{summary.den}
\alias{summary.den}
\title{Summary statistics for den}
\usage{
\method{summary}{den}(x, ...)
}
\arguments{
\item{x}{a den class object}
}
\value{
a list of markdown tables
}
\description{
Summary statistics for den
}
\examples{

}

