% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tags.to.sectors}
\alias{tags.to.sectors}
\title{Convert list of tags to sectors}
\usage{
tags.to.sectors(
  den,
  sector.tags,
  other = "Other",
  silent = FALSE,
  mutually.exclusive = FALSE,
  sector.membership = FALSE
)
}
\arguments{
\item{den}{a den class object}

\item{sector.tags}{a list of tags}

\item{other}{The other category, if FALSE, it is omitted}

\item{silent}{if FALSE the number of matched positions and affiliations is printed}

\item{mutually.exclusive}{if TRUE the produced sectors are mutually exclusive}

\item{sector.membership}{if TRUE a data.frame with a mutually exclusive sector memberships vector is returned}
}
\value{
a list of den objects
}
\description{
Convert list of tags to sectors
}
\examples{
data(den)
sectors        <- standard.sectors("Danish")
tags.to.sectors(den, sectors)
tags.to.sectors(den, sectors, mutually.exclusive = TRUE, sector.membership = TRUE)
}
