% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph creation.R
\name{two.mode}
\alias{two.mode}
\title{Create a two-mode network from an affiliation list}
\usage{
two.mode(den)
}
\arguments{
\item{den}{an affiliation list}
}
\description{
Create a two-mode network from an affiliation list
}

