% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{vertex.measures}
\alias{vertex.measures}
\title{Vertex centrality measures}
\usage{
vertex.measures(graph, mode = "all", order = 2.1)
}
\arguments{
\item{graph}{is an \link{igraph} network object}

\item{order}{is the maximum distance between two individuals for the reach
statistic}
}
\value{
a data.frame with a lot of descriptives
}
\description{
The classical centrality measures; degree, weighted degree, closenesss,
reach, betweenness and Burts constraint. This function is useful when you want to export results.
}
\examples{
data(pe13)
vertex.measures(net.elite)
}
