% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{weighted.adjacency.list}
\alias{weighted.adjacency.list}
\title{Create a list af weighted adjacency matrices}
\usage{
weighted.adjacency.list(spell.graph, start, end, to.distance = TRUE)
}
\arguments{
\item{spell.graph}{a spell.graph}

\item{start}{}

\item{end}{}

\item{to.distance}{if TRUE weights are recalculated with \link{distance.weight}, otherwise weighted months}
}
\description{
A more efficient way of creating time series of \link{weighted.graph}
}
