% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{weighted.graph}
\alias{weighted.graph}
\title{Title}
\usage{
weighted.graph(
  spell.graph,
  start,
  end,
  to.distance = TRUE,
  distance.weight = distance.weight,
  decay = decay,
  m = 38
)
}
\arguments{
\item{spell.graph}{}

\item{start}{}

\item{end}{}

\item{to.distance}{}

\item{distance.weight}{}

\item{decay}{}
}
\value{

}
\description{
Title
}
