% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph creation.R
\name{elite.network}
\alias{elite.network}
\title{Create a weighted network graph}
\usage{
elite.network(den, result = c("ind", "affil", "two.mode"), sigma = 14,
  base = exp(1))
}
\arguments{
\item{den}{a den class object, see \link{as.den}}

\item{result}{the type of returned graph}

\item{sigma}{the number of members in an affiliation above which all affiliations are weighted down}
}
\value{
a weighted graph
}
\description{
Create a weighted network graph
}
\examples{
data(den)
den.pol       <- has.tags(den, "Politics", result = "den")
elite.network(den.pol)
elite.network(den.pol, result = "affil")
elite.network(den.pol, result = "two.mode")
}
