% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliter.R
\docType{package}
\name{eliter-package}
\alias{eliter}
\alias{eliter-package}
\title{eliter: Contains the Danish Elite Network for 2012-2013 with analytical functions}
\description{
A set of tools for social network analysis with a focus on elites,
core identification and affiliation networks.
}
