% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{graph.from.spells}
\alias{graph.from.spells}
\title{Create an igraph object from position spells}
\usage{
graph.from.spells(den, diagonal = FALSE, minimum.duration = 1)
}
\arguments{
\item{den}{a den class object}

\item{diagonal}{if TRUE self ties, sometimes called loops, are included}

\item{minimum.duration}{the minimum edge duration in number of days}
}
\value{
an igraph object
}
\description{
Makes a graph from a den object with time spells e.g. a row looks like this;
NAME = Sam Samsen, AFFILIATION = Samson affiliates, PERSON_START = "2009-08-03", PERSON_END = "2009-12-02".
PERSON_START and PERSON_END has to be Date objects, see \link{as.Date} and \link{lubridate}.
}

