% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core identification.R
\name{k.shell}
\alias{k.shell}
\title{K-shell decomposition}
\usage{
k.shell(graph, start.level = 0, verbose = FALSE)
}
\arguments{
\item{graph}{a graph from igraph}

\item{start.level}{the decomposition process starts at this level or minimum degree}

\item{verbose}{if TRUE shows the progress}
}
\value{
a numeric vector
}
\description{
K-shell decomposition
}
