% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.den}
\alias{print.den}
\title{Print den objects}
\usage{
\method{print}{den}(x, ...)
}
\arguments{
\item{x}{a den-class object}

\item{...}{further arguments are ignored}
}
\value{
a markdown table with statistics
}
\description{
Essential statistics for the den-class. Mainly for interactive use, but results are returned in Rmarkdown format using kable and knitr for easy export.
}
\examples{
data(den)
as.den(den)
}

