% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{sectors.to.role}
\alias{sectors.to.role}
\title{Roles and position on the basis of sectors}
\usage{
sectors.to.role(
  den,
  list.dens,
  role = c("Chief executive", "Executive"),
  other = "Other",
  mutually.exclusive = FALSE
)
}
\arguments{
\item{den}{a den class object, see \link{as.den} with a ROLE variable}

\item{list.dens}{a list of den objects that corresponds to the sectors}

\item{role}{the set of roles}

\item{other}{either FALSE or the label for the "Other" category.}

\item{mutually.exclusive}{if TRUE sectors are mututally exclusive}
}
\value{
either a matrix or a named vector
}
\description{
sectors.to.role constructs a role or position variable on the basis of sectors.
If the sectors are mutually exclusive it constructs a single named vector.
If a person has roles in more than one sector it is the last sector in the list of sectors that counts.
}
\examples{
data(den)
den                 <- as.den(den)
list.dens           <- tags.to.sectors(den, standard.sectors("Danish"))
sectors.to.role(den, list.dens, mutually.exclusive = TRUE)
}
